refactor: updated config field to indexer
