+++ protocol-units/bridge/config/src/common/mod.rs
@@ -1,5 +1,5 @@
-pub mod db;
 pub mod eth;
+pub mod indexer;
 pub mod movement;
 pub mod testing;
